Fix package and rename response classes
