+++ app/src/main/java/org/ageage/eggplant/common/api/BookmarkService.kt
@@ -1,14 +1,14 @@
 package org.ageage.eggplant.common.api
 
 import io.reactivex.Observable
-import org.ageage.eggplant.bookmarks.BookmarkStarResponse
-import org.ageage.eggplant.common.api.response.BookmarkEntry
+import org.ageage.eggplant.common.api.response.BookmarkEntryResponse
+import org.ageage.eggplant.common.api.response.BookmarkStarResponse
 import retrofit2.http.GET
 import retrofit2.http.Query
 
 interface BookmarkService {
     @GET("/entry/jsonlite/")
-    fun bookmarkEntry(@Query("url") url: String): Observable<BookmarkEntry>
+    fun bookmarkEntry(@Query("url") url: String): Observable<BookmarkEntryResponse>
 
     @GET("/entry.json")
     fun startCount(@Query("uri") url: String): Observable<BookmarkStarResponse>
