+++ app/src/main/java/org/ageage/eggplant/common/api/response/BookmarkEntryResponse.kt
@@ -2,7 +2,7 @@ package org.ageage.eggplant.common.api.response
 
 import com.google.gson.annotations.SerializedName
 
-data class BookmarkEntry(
+data class BookmarkEntryResponse(
     val title: String,
     val count: Int,
     val entryUrl: String,
